Code cleanup (#23362)
* Code cleanup

* More cleanup

* Fix rename damage and other minor issues

* Lint, revert accidental changes

* Update external refs
