+++ operator/cmd/mesh/manifest-common.go
@@ -16,17 +16,12 @@ package mesh
 
 import (
 	"fmt"
-	"io"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 	"strings"
 
 	"github.com/ghodss/yaml"
 
 	"istio.io/api/operator/v1alpha1"
 	"istio.io/istio/operator/pkg/helm"
-	"istio.io/istio/operator/pkg/name"
 	"istio.io/istio/operator/pkg/tpath"
 	"istio.io/istio/operator/pkg/util"
 	"istio.io/istio/operator/pkg/util/clog"
@@ -98,101 +93,4 @@ func applyInstallFlagAlias(flags []string, charts string) []string {
 		flags = append(flags, fmt.Sprintf("installPackagePath=%s", charts))
 	}
 	return flags
-}
-
-type componentNameToListMap map[name.ComponentName][]name.ComponentName
-type componentTree map[name.ComponentName]interface{}
-
-var componentDependencies = componentNameToListMap{
-	name.PilotComponentName: {
-		name.PolicyComponentName,
-		name.TelemetryComponentName,
-		name.CNIComponentName,
-		name.IngressComponentName,
-		name.EgressComponentName,
-		name.AddonComponentName,
-	},
-	name.IstioBaseComponentName: {
-		name.PilotComponentName,
-	},
-}
-
-var installTree = make(componentTree)
-
-var dependencyWaitCh = make(map[name.ComponentName]chan struct{})
-
-func init() {
-	buildInstallTree()
-	for _, parent := range componentDependencies {
-		for _, child := range parent {
-			dependencyWaitCh[child] = make(chan struct{}, 1)
-		}
-	}
-}
-
-// RenderToDir writes manifests to a local filesystem directory tree.
-func RenderToDir(manifests name.ManifestMap, outputDir string, dryRun bool, l clog.Logger) error {
-	l.LogAndPrint("Component dependencies tree: \n%s", installTreeString())
-	l.LogAndPrint("Rendering manifests to output dir %s", outputDir)
-	return renderRecursive(manifests, installTree, outputDir, dryRun, l)
-}
-
-func renderRecursive(manifests name.ManifestMap, installTree componentTree, outputDir string, dryRun bool, l clog.Logger) error {
-	for k, v := range installTree {
-		componentName := string(k)
-		// In cases (like gateways) where multiple instances can exist, concatenate the manifests and apply as one.
-		ym := strings.Join(manifests[k], helm.YAMLSeparator)
-		l.LogAndPrint("Rendering: %s", componentName)
-		dirName := filepath.Join(outputDir, componentName)
-		if !dryRun {
-			if err := os.MkdirAll(dirName, os.ModePerm); err != nil {
-				return fmt.Errorf("could not create directory %s; %s", outputDir, err)
-			}
-		}
-		fname := filepath.Join(dirName, componentName) + ".yaml"
-		l.LogAndPrint("Writing manifest to %s", fname)
-		if !dryRun {
-			if err := ioutil.WriteFile(fname, []byte(ym), 0644); err != nil {
-				return fmt.Errorf("could not write manifest config; %s", err)
-			}
-		}
-
-		kt, ok := v.(componentTree)
-		if !ok {
-			// Leaf
-			return nil
-		}
-		if err := renderRecursive(manifests, kt, dirName, dryRun, l); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func buildInstallTree() {
-	// Starting with root, recursively insert each first level child into each node.
-	insertChildrenRecursive(name.IstioBaseComponentName, installTree, componentDependencies)
-}
-
-func insertChildrenRecursive(componentName name.ComponentName, tree componentTree, children componentNameToListMap) {
-	tree[componentName] = make(componentTree)
-	for _, child := range children[componentName] {
-		insertChildrenRecursive(child, tree[componentName].(componentTree), children)
-	}
-}
-
-func installTreeString() string {
-	var sb strings.Builder
-	buildInstallTreeString(name.IstioBaseComponentName, "", &sb)
-	return sb.String()
-}
-
-func buildInstallTreeString(componentName name.ComponentName, prefix string, sb io.StringWriter) {
-	_, _ = sb.WriteString(prefix + string(componentName) + "\n")
-	if _, ok := installTree[componentName].(componentTree); !ok {
-		return
-	}
-	for k := range installTree[componentName].(componentTree) {
-		buildInstallTreeString(k, prefix+"  ", sb)
-	}
 }
