+++ operator/cmd/mesh/manifest-generate.go
@@ -16,7 +16,9 @@ package mesh
 
 import (
 	"fmt"
+	"io/ioutil"
 	"os"
+	"path/filepath"
 	"sort"
 	"strings"
 
@@ -26,6 +28,7 @@ import (
 	"istio.io/api/operator/v1alpha1"
 	"istio.io/istio/operator/pkg/controlplane"
 	"istio.io/istio/operator/pkg/helm"
+	"istio.io/istio/operator/pkg/helmreconciler"
 	"istio.io/istio/operator/pkg/name"
 	"istio.io/istio/operator/pkg/translate"
 	"istio.io/istio/operator/pkg/util/clog"
@@ -49,9 +52,9 @@ type manifestGenerateArgs struct {
 
 func addManifestGenerateFlags(cmd *cobra.Command, args *manifestGenerateArgs) {
 	cmd.PersistentFlags().StringSliceVarP(&args.inFilename, "filename", "f", nil, filenameFlagHelpStr)
-	cmd.PersistentFlags().StringVarP(&args.outFilename, "output", "o", "", "Manifest output directory path")
-	cmd.PersistentFlags().StringArrayVarP(&args.set, "set", "s", nil, SetFlagHelpStr)
-	cmd.PersistentFlags().BoolVar(&args.force, "force", false, "Proceed even with validation errors")
+	cmd.PersistentFlags().StringVarP(&args.outFilename, "output", "o", "", "Manifest output directory path.")
+	cmd.PersistentFlags().StringArrayVarP(&args.set, "set", "s", nil, setFlagHelpStr)
+	cmd.PersistentFlags().BoolVar(&args.force, "force", false, "Proceed even with validation errors.")
 	cmd.PersistentFlags().StringVarP(&args.charts, "charts", "d", "", chartsFlagHelpStr)
 }
 
@@ -137,7 +140,7 @@ func GenManifests(inFilename []string, setOverlayYAML string, force bool,
 		return nil, nil, err
 	}
 
-	cp, err := controlplane.NewIstioOperator(mergedIOPS, t)
+	cp, err := controlplane.NewIstioControlPlane(mergedIOPS, t)
 	if err != nil {
 		return nil, nil, err
 	}
@@ -152,6 +155,7 @@ func GenManifests(inFilename []string, setOverlayYAML string, force bool,
 	return manifests, mergedIOPS, nil
 }
 
+// orderedManifests generates a list of manifests from the given map sorted by the map keys.
 func orderedManifests(mm name.ManifestMap) []string {
 	var keys, out []string
 	for k := range mm {
@@ -162,4 +166,43 @@ func orderedManifests(mm name.ManifestMap) []string {
 		out = append(out, strings.Join(mm[name.ComponentName(k)], helm.YAMLSeparator))
 	}
 	return out
+}
+
+// RenderToDir writes manifests to a local filesystem directory tree.
+func RenderToDir(manifests name.ManifestMap, outputDir string, dryRun bool, l clog.Logger) error {
+	l.LogAndPrint("Component dependencies tree: \n%s", helmreconciler.InstallTreeString())
+	l.LogAndPrint("Rendering manifests to output dir %s", outputDir)
+	return renderRecursive(manifests, helmreconciler.InstallTree, outputDir, dryRun, l)
+}
+
+func renderRecursive(manifests name.ManifestMap, installTree helmreconciler.ComponentTree, outputDir string, dryRun bool, l clog.Logger) error {
+	for k, v := range installTree {
+		componentName := string(k)
+		// In cases (like gateways) where multiple instances can exist, concatenate the manifests and apply as one.
+		ym := strings.Join(manifests[k], helm.YAMLSeparator)
+		l.LogAndPrint("Rendering: %s", componentName)
+		dirName := filepath.Join(outputDir, componentName)
+		if !dryRun {
+			if err := os.MkdirAll(dirName, os.ModePerm); err != nil {
+				return fmt.Errorf("could not create directory %s; %s", outputDir, err)
+			}
+		}
+		fname := filepath.Join(dirName, componentName) + ".yaml"
+		l.LogAndPrint("Writing manifest to %s", fname)
+		if !dryRun {
+			if err := ioutil.WriteFile(fname, []byte(ym), 0644); err != nil {
+				return fmt.Errorf("could not write manifest config; %s", err)
+			}
+		}
+
+		kt, ok := v.(helmreconciler.ComponentTree)
+		if !ok {
+			// Leaf
+			return nil
+		}
+		if err := renderRecursive(manifests, kt, dirName, dryRun, l); err != nil {
+			return err
+		}
+	}
+	return nil
 }
