+++ operator/cmd/mesh/operator-common.go
@@ -16,23 +16,14 @@ package mesh
 
 import (
 	"context"
-	"fmt"
-	"os"
 
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	v12 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/serializer"
 	"k8s.io/client-go/kubernetes"
-	"k8s.io/client-go/kubernetes/scheme"
 	_ "k8s.io/client-go/plugin/pkg/client/auth" // Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
-	"k8s.io/client-go/rest"
-	"k8s.io/client-go/tools/clientcmd"
 
 	"istio.io/istio/operator/pkg/helm"
 	"istio.io/istio/operator/pkg/name"
 	"istio.io/istio/operator/pkg/util"
-	"istio.io/istio/operator/pkg/util/clog"
 )
 
 type operatorCommonArgs struct {
@@ -54,12 +45,13 @@ const (
 	istioDefaultNamespace    = "istio-system"
 )
 
+// isControllerInstalled reports whether an operator deployment exists in the given namespace.
 func isControllerInstalled(cs kubernetes.Interface, operatorNamespace string) (bool, error) {
-	return DeploymentExists(cs, operatorNamespace, operatorResourceName)
+	return deploymentExists(cs, operatorNamespace, operatorResourceName)
 }
 
-// chartsRootDir, helmBaseDir, componentName, namespace string) (Template, TemplateRenderer, error) {
-func renderOperatorManifest(_ *rootArgs, ocArgs *operatorCommonArgs, _ clog.Logger) (string, string, error) {
+// renderOperatorManifest renders a manifest to install the operator with the given input arguments.
+func renderOperatorManifest(_ *rootArgs, ocArgs *operatorCommonArgs) (string, string, error) {
 	installPackagePath := snapshotInstallPackageDir
 	if ocArgs.charts != "" {
 		installPackagePath = ocArgs.charts
@@ -99,88 +91,11 @@ tag: {{.Tag}}
 	return vals, manifest, err
 }
 
-func CreateNamespace(cs kubernetes.Interface, namespace string) error {
-	if namespace == "" {
-		// Setup default namespace
-		namespace = istioDefaultNamespace
-	}
-
-	ns := &v1.Namespace{ObjectMeta: v12.ObjectMeta{
-		Name: namespace,
-		Labels: map[string]string{
-			"istio-injection": "disabled",
-		},
-	}}
-	_, err := cs.CoreV1().Namespaces().Create(context.TODO(), ns, v12.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return fmt.Errorf("failed to create namespace %v: %v", namespace, err)
-	}
-	return nil
-}
-
-func DeleteNamespace(cs kubernetes.Interface, namespace string) error {
-	return cs.CoreV1().Namespaces().Delete(context.TODO(), namespace, v12.DeleteOptions{})
-}
-
-func DeploymentExists(cs kubernetes.Interface, namespace, name string) (bool, error) {
+// deploymentExists returns true if the given deployment in the namespace exists.
+func deploymentExists(cs kubernetes.Interface, namespace, name string) (bool, error) {
 	d, err := cs.AppsV1().Deployments(namespace).Get(context.TODO(), name, v12.GetOptions{})
 	if err != nil {
 		return false, err
 	}
 	return d != nil, nil
-}
-
-func InitK8SRestClient(kubeconfig, kubeContext string) (*rest.Config, *kubernetes.Clientset, error) {
-	restConfig, err := defaultRestConfig(kubeconfig, kubeContext)
-	if err != nil {
-		return nil, nil, err
-	}
-	clientset, err := kubernetes.NewForConfig(restConfig)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return restConfig, clientset, nil
-}
-
-func defaultRestConfig(kubeconfig, kubeContext string) (*rest.Config, error) {
-	config, err := BuildClientConfig(kubeconfig, kubeContext)
-	if err != nil {
-		return nil, err
-	}
-	config.APIPath = "/api"
-	config.GroupVersion = &v1.SchemeGroupVersion
-	config.NegotiatedSerializer = serializer.WithoutConversionCodecFactory{CodecFactory: scheme.Codecs}
-	return config, nil
-}
-
-// BuildClientConfig is a helper function that builds client config from a kubeconfig filepath.
-// It overrides the current context with the one provided (empty to use default).
-//
-// This is a modified version of k8s.io/client-go/tools/clientcmd/BuildConfigFromFlags with the
-// difference that it loads default configs if not running in-cluster.
-func BuildClientConfig(kubeconfig, context string) (*rest.Config, error) {
-	if kubeconfig != "" {
-		info, err := os.Stat(kubeconfig)
-		if err != nil || info.Size() == 0 {
-			// If the specified kubeconfig doesn't exists / empty file / any other error
-			// from file stat, fall back to default
-			kubeconfig = ""
-		}
-	}
-
-	//Config loading rules:
-	// 1. kubeconfig if it not empty string
-	// 2. In cluster config if running in-cluster
-	// 3. Config(s) in KUBECONFIG environment variable
-	// 4. Use $HOME/.kube/config
-	loadingRules := clientcmd.NewDefaultClientConfigLoadingRules()
-	loadingRules.DefaultClientConfig = &clientcmd.DefaultClientConfig
-	loadingRules.ExplicitPath = kubeconfig
-	configOverrides := &clientcmd.ConfigOverrides{
-		ClusterDefaults: clientcmd.ClusterDefaults,
-		CurrentContext:  context,
-	}
-
-	return clientcmd.NewNonInteractiveDeferredLoadingClientConfig(loadingRules, configOverrides).ClientConfig()
 }
