+++ operator/cmd/mesh/operator-dump.go
@@ -55,8 +55,9 @@ func operatorDumpCmd(rootArgs *rootArgs, odArgs *operatorDumpArgs) *cobra.Comman
 		}}
 }
 
+// operatorDump dumps the manifest used to install the operator.
 func operatorDump(args *rootArgs, odArgs *operatorDumpArgs, l clog.Logger) {
-	_, mstr, err := renderOperatorManifest(args, &odArgs.common, l)
+	_, mstr, err := renderOperatorManifest(args, &odArgs.common)
 	if err != nil {
 		l.LogAndFatal(err)
 	}
