+++ operator/cmd/mesh/operator_test.go
@@ -78,7 +78,7 @@ func TestOperatorInit(t *testing.T) {
 	}
 
 	l := clog.NewConsoleLogger(os.Stdout, os.Stderr, installerScope)
-	_, gotYAML, err := renderOperatorManifest(rootArgs, &oiArgs.common, l)
+	_, gotYAML, err := renderOperatorManifest(rootArgs, &oiArgs.common)
 	if err != nil {
 		l.LogAndFatal(err)
 	}
