+++ operator/cmd/mesh/root.go
@@ -25,7 +25,7 @@ import (
 )
 
 const (
-	SetFlagHelpStr = `Override an IstioOperator value, e.g. to choose a profile
+	setFlagHelpStr = `Override an IstioOperator value, e.g. to choose a profile
 (--set profile=demo), enable or disable components (--set components.policy.enabled=true), or override Istio
 settings (--set values.grafana.enabled=true). See documentation for more info:
 https://istio.io/docs/reference/config/istio.operator.v1alpha12.pb/#IstioControlPlaneSpec`
@@ -43,15 +43,11 @@ This flag can be specified multiple times to overlay multiple files. Multiple fi
 type rootArgs struct {
 	// Dry run performs all steps except actually applying the manifests or creating output dirs/files.
 	dryRun bool
-	// Verbose controls whether additional debug output is displayed and logged.
-	verbose bool
 }
 
 func addFlags(cmd *cobra.Command, rootArgs *rootArgs) {
 	cmd.PersistentFlags().BoolVarP(&rootArgs.dryRun, "dry-run", "",
 		false, "Console/log output only, make no changes.")
-	cmd.PersistentFlags().BoolVarP(&rootArgs.verbose, "verbose", "",
-		false, "Verbose output.")
 }
 
 // GetRootCmd returns the root of the cobra command-tree.
