+++ operator/pkg/helmreconciler/prune.go
@@ -22,6 +22,8 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"istio.io/istio/operator/pkg/cache"
+	"istio.io/istio/operator/pkg/name"
 	"istio.io/istio/operator/pkg/object"
 	"istio.io/istio/operator/pkg/util"
 )
@@ -92,20 +94,22 @@ var (
 )
 
 // Prune removes any resources not specified in manifests generated by HelmReconciler h.
-func (h *HelmReconciler) Prune(manifests ChartManifestsMap) error {
+func (h *HelmReconciler) Prune(manifests name.ManifestMap) error {
 	var errs util.Errors
 	for cname, manifest := range manifests.Consolidated() {
-		errs = util.AppendErr(errs, h.PruneUnlistedResources(allObjectHashes(manifest), cname))
+		errs = util.AppendErr(errs, h.pruneUnlistedResources(object.AllObjectHashes(manifest), cname))
 	}
 	return errs.ToError()
 }
 
 // Delete removes all resources associated with componentName.
 func (h *HelmReconciler) DeleteComponent(componentName string) error {
-	return h.PruneUnlistedResources(map[string]bool{}, componentName)
+	return h.pruneUnlistedResources(map[string]bool{}, componentName)
 }
 
-func (h *HelmReconciler) PruneUnlistedResources(excluded map[string]bool, componentName string) error {
+// pruneUnlistedResources prunes any resources from the given component that are not in the excluded map. Resource
+// labels are used to identify the resources belonging to the component.
+func (h *HelmReconciler) pruneUnlistedResources(excluded map[string]bool, componentName string) error {
 	var errs util.Errors
 	for _, gvk := range append(namespacedResources, nonNamespacedResources...) {
 		objects := &unstructured.UnstructuredList{}
@@ -138,4 +142,14 @@ func (h *HelmReconciler) PruneUnlistedResources(excluded map[string]bool, compon
 		}
 	}
 	return errs.ToError()
+}
+
+// RemoveObject removes object with objHash in componentName from the object cache.
+func (h *HelmReconciler) removeFromObjectCache(componentName, objHash string) {
+	crHash, err := h.getCRHash(componentName)
+	if err != nil {
+		scope.Error(err.Error())
+	}
+	cache.RemoveObject(crHash, objHash)
+	scope.Infof("Removed object %s from Cache.", objHash)
 }
